Fixes #7, #8, #9 by improving our parsing rules & how we look for attachments
